Handle line breaks in a more convenient way.
This patch handles line breaks in a way which is more convenient for
textarea editing: single line breaks are rendered as line breaks (<br>)
instead of being ignored and requiring to put double trailing spaces to
achieve the same effect.

The reason behind this is that while it might be convenient to wrap
manually ones own lines to 78 chars when editing a text file in vi or
emacs, it doesn't make sense to do so in a linuxfr textarea, hence it
doesn't make sense to support this usage. Also, textareas can't be
configured to display trailing whitespace, so relying on them is quite
hazardous.

Previous art for this modification includes github.