Simple proposal to allow comment edition
This patch is intended as a thought-provoker for comment edition, not
necessarily the final end-all way to enable this feature. That said,
I think it could already be put in production and would be valuable.

More specifically:
- anyone can now edit its own comments (besides admins and moderators)
- the date of last modification is displayed at the bottom of the
  comment (I'm not a web developper so HTML and CSS changes may be
  suboptimal)
- there is no support for recording changes and history

Comment editing *could* evolve towards full wiki-like edition with
support for changes and history. I also think that would be rather
easy to do for someone familiar with the wiki page/revision
organization.

I belive the present patch already provides value and would improve
linuxfr.org as it is.