Use full links to the content in comments.
Before this patch, the links in the comment boxes were not relevant when
the comment was not presented directly on the related content. For
instance on the comment's own page, the [^] link led to a non-existent
local anchor and the '#' link led to the comment on the page itself.

This fixes https://linuxfr.org/suivi/liens-non-fonctionnels-sur-la-page-du-commentaire-et-la-liste-des-commentaires-dun-utilisateur