Display a link to the initial content of the comments when relevant.
When a comment is displayed out of context (i.e. in the user's comment
list) there is no immediate way to know the type and the title of the
content the comment responds to. Indeed, the comment title usually has
nothing to do with the content title.

This commit introduces a link to the associated content, including the
type and the title of the content. This link does not appear in
threaded views.

This fixes https://linuxfr.org/suivi/liste-des-commentaires-afficher-le-contenu-initial