Remove links to /node/<id>/comment/<id> in the comment blocks.
Since the "remove comment" button has been moved to the main _comment
template, there is no use left to the historic own page of the comment.
So let's just forgot it and don't promote it anymore.

The old anchor link ('#') has been removed and the link on the comment
title was replaced with the link which was previously used on the
anchor.