Display links to wikipedia without prefix
When a user wants to reference a wikipedia page, in another language
than french, or reference the wiktionary, he could use a prefix like 'en' or 'wikt'.
This works well, but the prefix is displayed in the result page.

This commit try to parse the given input, to check if there is prefix,
and in this case do not display the prefix.

This fixes
http://linuxfr.org/suivi/option-de-langue-pour-les-liens-wikipedia