Fix column colours not going to the bottom of the page.
Depending on whether the content column is smaller or bigger than the
sidebar, we ended up with columns changing colours at the middle of the
page.

This commit emulates full height columns by using a custom gradient on
the column container (the <body> tag). This fixes the graphical glitch
on recent enough browsers while not altering anything on older ones.

See http://caniuse.com/#feat=background-img-opts for supported browsers.