CSS: Simplify the header bar to avoid absolute positioning.
Since we now rely on the natural flow instead of absolute positions for
the sidebar and main contents, we have the absolute certainty that there
won't be any gap even if the user forces a bigger font. Also changes
won't have to be reflected at other places anymore.

It causes no aesthetical change at all.