change size limit on board messages from 500 to 65536
The sanitization code would cut board messages at 500 characters. This
is much too little: it naturally happens that people write messages of
over 500 characters to explain their point of view -- eg. when
collaborating on the writing of a news item.

I left a limit in place, namely 65536, to avoid flooding attacks on
the server. This means that the upper limit on space usage on the
server is around 100Kio, which is less than uploaded images.