boards: Add norloge attribute to time tags to split norloge format and text display
Before, the data inside <time> was used as norloge format and as text to display
to user.

Now, the data inside <time> is only used as text to display.

Final norloge format is defined according to board and box context by the Ruby
code inside app/helpers/board_helper.rb and config/initializers/time_formats.rb.

As the content of config is not reachable from JavaScript code, we add an
attribute "norloge" with correctly formatted text according to the context
(board and box).