<span> tag aroung "Posted by", "Edited by" and "Moderated by" sections.
This commit adds a dedicated <span> tag around each of the named-above sections
so they can be easily identified and selected, then inserts a carriage return
between each of them in the HTML code.

This reproduces the default behavour (having a space between a point "." and
the first word of the next sentence), but also enables the CSS to emulate a
line break if needed.

This was primarily done for "Steelblue" CSS, which does on the enlighted news
at top of the front page ("phare" section).

EDIT: commit amended because class "edited_by" was already in use.
Class names "posted_by_spanblock", "edited_by_spanblock" and
"moderated_by_spanblock" chosen instead.