Adding a class vote-for or vote-against on comment's "li" element
Mark a whole comment as upvoted or downvoted through a class on the
comment's root element (which is a list item inside the whole comment
tree).

This can be used to enlight the whole box or to decorate a single element
inside through CSS selector, but it also can be used to easily implement
additional features such as automaticaly hiding disapproved comments
regardless of their score. These feature may be written in a global scope
inside the server code, proposed to users as Javascript extensions, or
offered as third-party components through GreaseMonkey scripts or such.

Implemented for CSS "Steelblue" as part of pull-request commits.