Add div#comments when composing a new comment
Wrap the comment section (including the one we reply to)
inside a div#comments on the composing page.

This applies to both a reply to an article (news/diary/forum entry)
or another comment, which are two distinct cases.

CSS such as Steelblue distinguish these two cases but don't rely on the
code to do so. However, they still need a well-placed div#comments box.