Adding "div.sideforms" around page bottom forms on stylesheet selection page.
Addition of a division named "div.sideforms" to allow URL box and file
selector to properly fit on the right of the CSS snapshot gallery.

The gallery is designed to occupy either 60% of the main container OR a
maximum width limit of 520 pixels (image native dimension 400×400 plus some
margins), whichever is the smalest, leaving the rest to text and notices.
The gallery is implemented as a left floating box, which can be wrapped only
by the main stream, that is plain text or at least an inline box (inline-block
or inline-flex).

We thus need an invisible wrapper so we can treat the content as regular boxes,
to draw a frame and do some item alignments inside. This can't be done with CSS
selectors only. Note that declaring <form> tags as inline blocks do work, but
they would no longer expand to fit the whole of the remaining width on the right.