Bug fix on new forum entry page : some tags where outside their container
Tags <article>, <form#new_post> and <dev.markdown_cheatse> where outside
the <main#contents> container as it was supposed to be, and as it is on
other similar pages (for instance, posting a new diary).

This was due to a lack of a proper indentation. This patch fixes this.
As usual, no regression induced with the main CSS.