Adding a class ".container" to section#container and section#redaction
This basically affects almost all the pages of the server, but induces
no regression, since neither of these tags had classes so far.

We mainly need this because there's one page that is not contained inside
a section#container tag: the news redaction page (that is, the most important
one), but it stills contains the articles that will be published.

Since there's over 200 differents statements affecting articles and other
"contained" contents, and considering the facts that each of these statements
are very often multiplied to match pseudo classes and such, and some of them
are either exceptions or context-specific stuff (e.g. #phare vs. #main), they
can't be easily wrapped inside scss brackets { } (which should be avoided
anyway when there's too much clauses over them. We have the problem elsewhere
with public pages supposed to have a dark background).

Another "bad" solution would consist of duplicating all of them in another
specific section decicated to #redaction. We don't want that neither.

The best solution remains the addition of a specific class that can easily
apply to any other "oddity" such as this one. This hack moreover is a very
small and simple one, and can be reverted without any difficulty.