Add odd/even parity classes on last table of user statistics
This last table is very hard to discriminate through CSS selectors,
because it has no specific classes set on the rows themselves, and
because each slice don't have a fixed number of rows neither.

Something much better would have been to wrap each slice inside its
own tbody block, but breaking the table cell lineage was prone to
induce regressions on other CSS.

If needed and appreciated, we'll do that as a specific task when this
submission is considered stable.