Missing wrapping tags on "users-comments" page.
Missing "nav.toolbox" and "div#comments" around appropriate
sections, and class "thread" to ul list on user page when choosing
"Accès aux derniers : . commentaires".

Even though it surely looks like a bug, this is not quite visible
with default CSS but Steelblue definitely needs them. These additions
cause no regression with standard CSS.

   ----

Please note that the %style block is here to make page selectors reappear:
They're hidden on purpose by RonRonnement to mask the one at top of the
page (line 932), and next selector is supposed to spot the last one to
make it reappear. Unfortunately, it doesn't work on this particular page,
but nobody ever noticed especially because those malformed tags were
never matched by the CSS.

We simply need to correct this in the default CSS (one row to add just after
the line quoted above) but this will be the point of another further patch
to come. This particular revision is there to show that everything's still
working fine without starting to modify the default CSS.