Adding "moderation_page" class to related pages's <body>
This patch adds a class named "moderation_page" among the
existing ones to the <body> tag of pages which controller
path begins with "moderation", making them easier to spot
by CSSes.

It's not only convenient (it would widely simplify the CSS
code) but it's also necessary on pages such as #news-index
and #polls-index, which can't be discriminated from the
outside to know if they're the common user pages or the
moderation ones.

As regards #news-index in particular, the two pages are
totally different. The moderation page contains everything
but article boxes, and also offers special services such as
a chat board. The rest is formed of plain text and html
links that thus should be displayed on a light-background
page, not a dark one.