Conditional class "threads" on UL depending on its parent
Add a class "thread" to a comment's UL inside the _preview.haml
file to be automaticaly rendered by the "render" statement depending
on its context.

We want to set it when the rendered comment is either a new comment
or an existing one being edited, thus in both cases being a top of
an thread excerpt. However, when the comment is a direct answer to
another comment, we now render it as an idented extension of the
first one, just like it would be inside the whole thread in the
article's page.

This still causes no regression of the original CSS.