Update footer nodes to be more responsive
To correctly display download link on mobile screens, this solution
had to remove the float:right CSS option and use more modern CSS grids.

So, on big screen, we display footer on two lines (one with
almost everything except vote div, the other with the vote div).

On small screen, we in one column and so we have one line by
functionnality.

The downside is the display with Internet Explorer is broken (download
links hidde the read it button).