chat_n_edit is a larger layout and is splitted on 3 parts
1. left-part is the "metadata" part with summary and links (will come
later)
2. middle-part is the "edition" part where the users work
3. right-part is the "toolpanel" with chat, revision history and some details

In the middle-part, a topbar is added with back button, last revision
summary and attendee list.

For the last revision summary, we've added to the Push message the
formatted date of revision creation. That's used for toolpanel and
revision history.

Note, that chat_n_edit layout will be completly broken for contrib CSS
as it has now 3 columns and new ids.