Fix the daily number of votes
The daily number of votes is computed every day, via a crontab. It
updates the accounts, and the account model needs to be valid for the
change to be applied. But, there was a validator on the length of
uploaded_stylesheet, and when a user has uploaded a stylesheet, the
uplodaded_stylesheet field was a mounter uploaded instance, not a
string, and its length was larger than the allower 255 characters. As
the uploaded_stylesheet in database is generated by the uploader (code
on the server), and not controlled from client side, we can safely
remove the validation on the length, and it should fix the issue.