add /comments/latest route
This page will be useful for moderators to retrieve latest comments.

The limit is based on current date to limit number of comments retrieved
from the database. Only comments from the last 7 days will be available.

As it's the first page to index all comments with criteria only made on
created_at property, I added an index to the comments sql table.

As LinuxFr has between 1500 to 2000 comments each 15 days, I've decided
to paginate to 50 comments. As the number of comments in the atom feed is
linked to this pagination setting, I wanted it to be big enough so RSS readers
can take all comments with around 4 requests by day.