keep chosen license when a diary is converted to a news
This was the original intention of commit 6a8984b526dbb81f785e5c61d6435edf47902ac2
on March 2011. At that moment, the default value for `cc_licensed` on nodes
where `false` and we only needed to set it to true if the diary
license where the CC one.

Although, on December 2012, we changed the default value to `false` with
commit 7c6058f512bf004305eea2655d50bd1802b084e9 and it broke the diary
converter algorithm.

This fix allows to be not any more linked to the current default value
on the `cc_licensed` node as we always enfoce the new news to have the
same license.