notify node owner for new answer, add all content to dashboard
Notify owner of any type of content for new answers (in existing thread or
new thread).

Add a dashboard showing latest 10 contents published of any type, so the
author can follow easily new threads and comments on his/her content.

For small screends, the dashboards are scrollable to avoid them to break
the application layout.

Dashboard titles are aligned with content (aligned on right for numbers
and left for the rest).