Do not assume contributors of a news item are all males.
When there are more than 3 contributors to a news item, this is
initially displayed as “Édité par X contributeurs”, regardless of the
gender of the contributors.

This is consistent with official French grammar rules, but not
consistent with LinuxFr.org’s stated policy that masculine forms shall
not be used as a neutral form.

Therefore, we replace the male-specific “contributors” par the
gender-neutral “personnes”.