Do not assume developers are males.
Replace the text of the links pointing to /developer from “API pour les
développeurs” to “API pour les développeurs et développeuses”.

On that /developer page, similarly do not assume all those who would
want to use the API would necessarily be males.

When referring to a hypothetical user, since

* it would be way too cumbersome to systematically use a dual-gender
  form (e.g. “l’utilisateur ou l’utilisatrice”);

* “personne” is a bit too vague;

* I have a limited imagination and can’t think of another solution;

we’ll instead assume the hypothetical user is a she, and therefore
always use “utilisatrice”. This is consistent with information
theorists’ long-standing tradition of always assuming the first user
trying to communicate is a woman named Alice. :)