Include the Mastodon URL in a link element on user page (#2046)
We add a `<link rel="me" href="https://mastodon.example/@me">` URL in
the `<head>` when the user configures their Mastodon URL.

We only accept URLs (not `@me@mastodon.example`-style identifiers) for
the sake of simplicity, because it's both the most universal format and
the one required for the "verification" feature. A placeholder is used
to convey that fact.

For the record, at this time Mastodon requires a `<a>` or `<link>`
element with `rel="me"` and an `href` attribute pointing to the
`https://` URL of the Mastodon account. I opted for the `<link>` element
because it didn't involve modifying the web interface any further. See
https://docs.joinmastodon.org/user/profile/#verification for details.

https://linuxfr.org/suivi/faire-valider-sa-page-perso-linuxfr-org-sur-mastodon