RonRonnement CSS: Fix small discontinuity in the left sidebar
Commit 88aa3c387cd3c42d34b9b91c33d5a591a0b5bcdf included a small
computation mistake: the border should be taken into account in the
calculations.

Also, while it looks correct in Chrome, for some reason Firefox does not
properly position the linear gradient origin when the background is set
to `repeat`. Funnily it doesn't exhibit the same behaviour when using
the built-in screenshot tool. We add `no-repeat` to the background
specification to fix this.

We also remove the backgrond of the sidebar as it is useless since it is
the same colour as the body background, but it makes the miscalculation
obvious should one happen again.