Fix validating the URL of user home sites
Home site validation used the URL regexp, but didn't anchor it. Because
of this it was possible to add text before or after the URL and have it
considered valid. For instance, this was considered a valid homesite
value: "lorem ipsum https://example.org/#foo#bar dolor sit amet"

It is possible to fix it by anchoring the regexp, but IMO it is wiser to
use the same URL validator used at other places in the site. Hence this
commit replaces this validation with the `http_url` validator used for
news links and bookmarks.

https://linuxfr.org/suivi/impossible-de-mettre-un-lien-vers-un-salon-matrix-dans-les-liens-d-une-depeche#comment-1911550