fix URI validation, and btw multiple sharp signs in URIs
Fix validating the URL of user home sites

Home site validation used the URL regexp, but didn't anchor it. Because
of this it was possible to add text before or after the URL and have it
considered valid. For instance, this was considered a valid homesite
value: "lorem ipsum https://example.org/#foo#bar dolor sit amet"

It is possible to fix it by anchoring the regexp, but IMO it is wiser to
use the same URL validator used at other places in the site. Hence this
commit replaces this validation with the `http_url` validator used for
news links and bookmarks.

https://linuxfr.org/suivi/impossible-de-mettre-un-lien-vers-un-salon-matrix-dans-les-liens-d-une-depeche#comment-1911550

user: use the new URI validator instead of old HttpUrlValidator

remove comment reference to old HttpUrlValidator

adds after_validation method to UriValidator

This is used to revert the workaround applied on fragment part of the
URI to accept multiple sharp sign in the URI.

This decoding of sharp sign in the fragment part allows to keep visual match
with the user input.

bookmark controllers use the bookmark model validator to validate link

So the bookmark model is the only leader the validation of the link.

Thus the protocol validation is moved to the model too.