store date when account is last seen
This date will be used to drop personal user data if the account is
inactive for a certain period.

The automatic personal user data clean up has been announced here:
https://linuxfr.org/news/regles-de-perennite-des-comptes-linuxfr-org-et-donnees-a-caractere-personnel

This commit only add a column on Account and do not add scheduled jobs
to apply deletion.

To avoid to update database too often and to avoid to keep too accurate
user personal information, only the date will be saved on database.

Suivi request to add this information into database:
https://linuxfr.org/suivi/pouvoir-determiner-si-un-compte-a-eu-de-l-activite-recemment