Fix overlay totoz positioning
Positioning was previously using the jQuery offset() method to compute
the position. This method returns the absolute position of the element
relative to the whole document, but the values are then used to set the
absolute position (top and left CSS properties) of the totoz element.
These properties apply based on the closest positioned parent, which in
this case is the <body> element.

Due to this, when the document is larger than the max-width of the
<body> element (150ex) the overlay image was being pushed further right
by half the difference. For instance, if the window is 1600px wide, and
the body reaches its max width at 1213px in my case, the overlay is
being displayed (1600-1213)/2 = 193px to the right of the intended
position.

We fix this issue by using the jQuery position() instead, which
according to the documentation returns the absolute position of the
element relative to the offset parent, in this case the <body> element.
We also explicitly position the totoz below the [:...] anchor.