docker: Fix creating linuxfr.org using docker-compose
This commit fixes errors related to the recent-ish archival of Debian
Stretch. Indeed, the linuxfr.org container is still based on Debian
Stretch but this version of the distribution was archived when Bookworm
was released. This implies a change in the APT sources but neither the
debian:stretch-slim nor this repository were updated accordingly.

The errors fixed are:

 - we replace the `deb.debian.org` domain with `archive.debian.org`,
   where the archived repositories are located;
 - we downgrade `zlib1g` forcefully because the version which is already
   installed in the upstream container is newer than the one we can find
   in the archived repositories, and it caused a dependency issue when
   trying to install the `zlib1g-dev` package.