Remove CoffeeScript, use plain Javascript instead
CoffeeScript is (mostly) dead these days and, while it was advertised
with Rails 5 and older, it is not supported anymore by newer version of
Rails. Using plain Javascript might also make it easier for new
contributors to fix issues as nobody is familiar with CoffeeScript these
days.

In this commit we run `decaffeinate` on all the `.coffee` files and
nothing else.

See https://decaffeinate-project.org/

Suivi #3034