linuxfr.org container uses Debian Bullseye and backports container definition from rails7
The linuxfr.org container wasn't anymore buildable due to `npm ci` which
were stuck. So, even if the prod still use Debian Stretch, we had to
upgrade the development environment to Debian Bullseye.

Took the opportunity to backport container definition improvements from
rails7. Especially, containers are run without root user and there are
healthchecks which can help to better see the state of the containers.