validates lang selection for bookmarks and links
As title and link are both mandatory, we set the lang mandatory too.

The "xx" language code is removed from the valid list and blank is
refused.

The "xx" value was used as default value before we enable "blank" in
form. It is not really useful now except for links and bookmarks which
are already using this value.